dev to prod

+++ packages/editor/src/screens/DocumentWindow/DocumentTable.tsx
@@ -31,7 +31,8 @@ import {
 import { DocumentData, columns } from './document'
 import styles from './index.module.scss'
 import DocumentModal from './DocumentModal'
-
+import { useTreeData } from "../../../../core/client/src/contexts/TreeDataProvider"
+import DocContentModal from './DocContentModal'
 /**
  * GlobalFilter component for applying search filter on the whole table.
  * @param {{ globalFilter: any, setGlobalFilter: Function }} param0
@@ -82,10 +83,13 @@ function DocumentTable({ documents, updateCallback }) {
   const config = useConfig()
   const globalConfig = useSelector((state: any) => state.globalConfig)
   const token = globalConfig?.token
-
+  const [document , setDocument] = useState(null)
+  const [contentModal, setContentModal] = useState(false)
   const [anchorEl, setAnchorEl] = useState(null)
-  const [selectedRow, setSelectedRow] = useState(null)
+  // todo better typing here for the row
+  const [selectedRow, setSelectedRow] = useState<any>(null)
   const [currentPage, setCurrentPage] = useState(0)
+  const { setDocState,setToDelete ,  openDoc } = useTreeData();
 
   const handleActionClick = (document, row) => {
     setAnchorEl(document.currentTarget)
@@ -148,12 +152,14 @@ function DocumentTable({ documents, updateCallback }) {
   // Initialize the table with hooks
   const { page, flatRows, pageOptions, gotoPage, setGlobalFilter, state: { sortBy, globalFilter },
     setSortBy } =
-    useTable(
+    useTable<any>(
       {
         columns: defaultColumns,
         data: documents,
-        initialState : {
-          pageIndex: currentPage 
+        initialState: {
+          // todo we need to pass a proper generic into the useTable hook to fix this type error
+          // @ts-ignore
+          pageIndex: currentPage
         }
       },
       useFilters,
@@ -192,14 +198,18 @@ function DocumentTable({ documents, updateCallback }) {
 
   // Handle document deletion
   const handleDocumentDelete = async (document: any) => {
-    console.log('deleting document', document)
+    if (!selectedRow) return
     const isDeleted = await fetch(`${API_ROOT_URL}/documents/${selectedRow.id}`, {
       method: 'DELETE',
       headers: {
         Authorization: `Bearer ${token}`,
       },
     })
-    if (isDeleted) enqueueSnackbar('document deleted', { variant: 'success' })
+    if (isDeleted) {
+      enqueueSnackbar('document deleted', { variant: 'success' })
+      setToDelete(selectedRow.id)
+      setDocState(true);
+    }
     else enqueueSnackbar('Error deleting document', { variant: 'error' })
 
     if (page.length === 1) {
@@ -228,55 +238,82 @@ function DocumentTable({ documents, updateCallback }) {
     projectId: '',
     date: new Date().toISOString(),
     embedding: '',
+    files: []
   })
-// Handle save action
-const handleSave = async (selectedModel) => {
-  // call documents endpoint
-  const result = await fetch(`${API_ROOT_URL}/documents`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: `Bearer ${token}`,
-    },
-    body: JSON.stringify({
-      date: new Date().toISOString(),
-      ...newDocument,
-      projectId: config.projectId,
-      modelName: selectedModel.model,
-      secrets: localStorage.getItem('secrets'),
-    }),
-  });
-  // Check if the save operation was successful
-  if (result.ok) {
-    // Reset newDocument
-    setNewDocument({
-      type: '',
-      content: '',
-      projectId: '',
-      date: '',
-      embedding: '',
+  // Handle save action
+  const handleSave = async (selectedModel) => {
+    const { files, ...body } = newDocument
+    // call documents endpoint
+
+    const formData = new FormData();
+    formData.append('date', new Date().toISOString())
+    formData.append('projectId', config.projectId)
+    formData.append('modelName', selectedModel.model)
+    formData.append('secrets', localStorage.getItem('secrets') || '')
+    formData.append('type', body.type)
+    formData.append('content', body.content)
+    for (const file of files as File[]) {
+      formData.append('files', file, file.name)
+    }
+    const result = await fetch(`${API_ROOT_URL}/documents`, {
+      method: 'POST',
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+      body: formData,
     });
-    enqueueSnackbar('Document saved successfully', { variant: 'success' });
+    // Check if the save operation was successful
+    if (result.ok) {
+      // Reset newDocument
+      setNewDocument({
+        type: '',
+        content: '',
+        projectId: '',
+        date: '',
+        embedding: '',
+        files: []
+      });
+      enqueueSnackbar('Document saved successfully', { variant: 'success' });
 
 
-    // Close the modal by setting createMode to false after a delay
-    setTimeout(() => {
-      setCreateMode(false);
-    }, 2000);
+      // Close the modal by setting createMode to false after a delay
+      setTimeout(() => {
+        setDocState(true);
+        setCreateMode(false);
+      }, 2000);
 
-    // Trigger the updateCallback function to update the table after a delay
-   
+      // Trigger the updateCallback function to update the table after a delay
+
       updateCallback();
-    
-  } else {
-    enqueueSnackbar('Error saving document', { variant: 'error' });
-  }
-};
+
+    } else {
+      enqueueSnackbar('Error saving document', { variant: 'error' });
+    }
+  };
 
   // Show create modal
   const showCreateModal = () => {
     setCreateMode(true)
   }
+
+  const handleFindDoc = (doc) => {
+   //fetch the document 
+    fetch(`${API_ROOT_URL}/documents/${doc}`, {
+      method: 'GET',
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    })
+      .then(res => res.json())
+      .then((res) => {
+        setDocument(res.content)
+        setContentModal(true)
+      })
+      .catch((err) => {
+        console.log(err)
+      })
+  }
+
 
   // trigger updateCallback when createMode changes
   useEffect(() => {
@@ -284,6 +321,17 @@ const handleSave = async (selectedModel) => {
       updateCallback();
     }
   }, [createMode]);
+
+  useEffect(() => {
+    if (openDoc) {
+      handleFindDoc(openDoc)
+      console.log(openDoc);
+      
+    }
+  }, [openDoc])
+
+  console.log(document);
+
   return (
     <>{createMode && (
       <DocumentModal
@@ -294,6 +342,13 @@ const handleSave = async (selectedModel) => {
         providerList={filteredProviders}
       />
     )}
+      {contentModal && document  && (
+        <DocContentModal
+          contentModal={contentModal}
+          setContentModal={setContentModal}
+          document={document}
+        />
+      )}
       <Container className={styles.container} classes={{ root: styles.root }}>
         <Stack spacing={2} style={{ padding: '1rem', background: '#272727' }}>
           <div className={styles.flex}>
